Makefile: Remove incorrect and suprefluous KDIR variable definition
The KDIR variable assignment gets overwritten by the KDIR=... in the
make invocation started from dkms. So it gets never used.
(See:
https://stackoverflow.com/questions/2826029/passing-additional-variables-from-command-line-to-make
)

Even if it would be used it would lead to compilation errors, since KDIR
has to be defined by dkms (when it "cross-compiles" for other kernels), not by
uname with our current kernel